requirements.txt
streamlit
pandas
matplotlib
plotly
numpy
scikit-learn
pymongo
google-generativeai
